requirements.txt
streamlit
typing-extensions
langchain
openai
pydantic==1.10.8
python-dotenv
watchdog
wikipedia